+++ Core/Opcodes/AddOpcode.cs
@@ -55,7 +55,7 @@
 			}
 
 			// Now yes, do it
-			int sum = ( (int) op1.LiteralValue.Value ) + ( (int) op2.LiteralValue.Value );
+			long sum = ( (long) op1.LiteralValue.Value ) + ( (long) op2.LiteralValue.Value );
 
 			// Store in the temp vble and end
 			Variable result = new TempVariable( new IntLiteral( this.Machine, sum ) );
